+++ MangAdventure/__init__.py
@@ -1,5 +1,5 @@
 """A simple manga hosting CMS written in Django."""
 
 __license__ = 'MIT'
-__version__ = '0.6.0'
+__version__ = '0.6.1'
 __author__ = 'evangelos-ch, ObserverOfTime'