+++ MangAdventure/search.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, List, NamedTuple, Tuple
 
 from django.db.models.query import Q
 
-from reader.models import ArtistAlias, AuthorAlias, Series, SeriesAlias
+from reader.models import Series
 
 if TYPE_CHECKING:  # pragma: no cover
     from django.db.models.query import QuerySet
@@ -40,7 +40,7 @@ def parse(request: 'HttpRequest') -> _SearchParams:
     return _SearchParams(
         query=request.GET.get('q', ''),
         author=request.GET.get('author', ''),
-        status=request.GET.get('status', 'any'),
+        status=request.GET.get('status', 'any').lower(),
         categories=(
             [c.lower() for c in categories if len(c) and c[0] != '-'],
             [c[1:].lower() for c in categories if len(c) and c[0] == '-']
@@ -48,13 +48,13 @@ def parse(request: 'HttpRequest') -> _SearchParams:
     )
 
 
-def qsfilter(params: _SearchParams) -> Tuple[Q, List[str]]:
+def qsfilter(params: _SearchParams) -> Q:
     """
     Create a `queryset filter`_ from the given search parameters.
 
     :param params: A :obj:`~collection.namedtuple` of parameters.
 
-    :return: A tuple containing the filter and the excluded categories.
+    :return: The created queryset filter.
 
     .. _`queryset filter`:
         https://docs.djangoproject.com/en/3.0/
@@ -62,32 +62,27 @@ def qsfilter(params: _SearchParams) -> Tuple[Q, List[str]]:
     """
     filters = Q()
     if params.query:
-        filters = Q(title__icontains=params.query)
-        aliases = SeriesAlias.objects.filter(
-            alias__icontains=params.query
+        filters = (
+            Q(title__icontains=params.query) |
+            Q(aliases__alias__icontains=params.query)
         )
-        if len(aliases):
-            filters |= Q(aliases__in=aliases)
     if params.author:
-        q = Q(authors__name__icontains=params.author) | \
-            Q(artists__name__icontains=params.author)
-        authors = AuthorAlias.objects.filter(
-            alias__icontains=params.author
+        filters &= (
+            Q(authors__name__icontains=params.author) |
+            Q(artists__name__icontains=params.author) |
+            Q(authors__aliases__alias__icontains=params.author) |
+            Q(artists__aliases__alias__icontains=params.author)
         )
-        if len(authors):
-            q |= Q(authors__pk__in=authors)
-        artists = ArtistAlias.objects.filter(
-            alias__icontains=params.author
-        )
-        if len(artists):
-            q |= Q(artists__pk__in=artists)
-        filters &= q
-    if params.status != 'any':
-        filters &= Q(completed=(params.status == 'completed'))
-    categories = params.categories
-    if params.categories[0]:
-        filters &= Q(categories__in=categories[0])
-    return filters, categories[1]
+    if params.status == 'completed':
+        filters &= Q(completed=True)
+    elif params.status == 'ongoing':
+        filters &= Q(completed=False)
+    included, excluded = params.categories
+    if excluded:
+        filters &= ~Q(categories__in=excluded)
+    if included:
+        filters &= Q(categories__in=included)
+    return filters
 
 
 def query(params: _SearchParams) -> 'QuerySet':
@@ -98,9 +93,7 @@ def query(params: _SearchParams) -> 'QuerySet':
 
     :return: A queryset of series matching the given parameters.
     """
-    filters, exclude = qsfilter(params)
-    return Series.objects.filter(filters) \
-        .distinct().exclude(categories__in=exclude)
+    return Series.objects.filter(qsfilter(params)).distinct()
 
 
 def get_response(request: 'HttpRequest') -> 'QuerySet':